+++ click_extended/decorators/transform/__init__.py
@@ -1,6 +1,7 @@
 """Initialization file for the `click_extended.decorators.transform` module."""
 
 from click_extended.decorators.transform.apply import apply
+from click_extended.decorators.transform.strip import lstrip, rstrip, strip
 from click_extended.decorators.transform.to_case import (
     to_camel_case,
     to_dot_case,
@@ -24,6 +25,9 @@ from click_extended.decorators.transform.to_timestamp import to_timestamp
 
 __all__ = [
     "apply",
+    "lstrip",
+    "rstrip",
+    "strip",
     "to_camel_case",
     "to_dot_case",
     "to_flat_case",